Split line tests from test_geometry.py
// edited by skirpichev
